backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
dj-database-url==3.0.1
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
frozendict==2.4.7
gunicorn==23.0.0
idna==3.11
multitasking==0.0.12
numpy==2.0.2
packaging==26.0
pandas==2.3.3
peewee==4.0.1
platformdirs==4.4.0
protobuf==6.33.5
psycopg2-binary==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.11.0
yfinance==1.2.0
